requirementsD3.txt
arabic-reshaper==2.1.3
asgiref==3.6.0
astroid==2.3.3
colorama==0.4.3
configparser==5.3.0
dj-database-url==0.5.0
Django==3.2
django-crum==0.7.9
django-filter==22.1
django-userforeignkey==0.3.0
djangorestframework==3.14.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
protobuf==3.20.1
psycopg2==2.9.5
pylint==2.4.4
PyMySQL==1.0.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.3
pytz==2021.3
reportlab==3.6.6
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.11.1
xhtml2pdf==0.2.5